Removed Settings panel in favor of new shared Settings modal
