+++ src/devtools/views/Components/Components.js
@@ -7,27 +7,32 @@ import { InspectedElementContextController } from './InspectedElementContext';
 import { OwnersListContextController } from './OwnersListContext';
 import portaledContent from '../portaledContent';
 import { ModalDialog } from '../ModalDialog';
+import SettingsModal from 'src/devtools/views/Settings/SettingsModal';
+import { SettingsModalContextController } from 'src/devtools/views/Settings/SettingsModalContext';
 
 import styles from './Components.css';
 
 function Components(_: {||}) {
   // TODO Flex wrappers below should be user resizable.
   return (
-    <OwnersListContextController>
-      <InspectedElementContextController>
-        <div className={styles.Components}>
-          <div className={styles.TreeWrapper}>
-            <Tree />
+    <SettingsModalContextController>
+      <OwnersListContextController>
+        <InspectedElementContextController>
+          <div className={styles.Components}>
+            <div className={styles.TreeWrapper}>
+              <Tree />
+            </div>
+            <div className={styles.SelectedElementWrapper}>
+              <Suspense fallback={<Loading />}>
+                <SelectedElement />
+              </Suspense>
+            </div>
+            <ModalDialog />
+            <SettingsModal defaultTabID="components" />
           </div>
-          <div className={styles.SelectedElementWrapper}>
-            <Suspense fallback={<Loading />}>
-              <SelectedElement />
-            </Suspense>
-          </div>
-          <ModalDialog />
-        </div>
-      </InspectedElementContextController>
-    </OwnersListContextController>
+        </InspectedElementContextController>
+      </OwnersListContextController>
+    </SettingsModalContextController>
   );
 }
