--- src/devtools/views/Components/ToggleComponentFiltersModalButton.css
@@ -1,8 +1,0 @@
-.Wrapper {
-  display: flex;
-}
-
-.Label {
-  font-size: var(--font-size-sans-small);
-  font-family: var(--font-family-sans);
-}
