+++ src/devtools/views/Profiler/Profiler.css
@@ -1,6 +1,7 @@
 .Profiler {
   width: 100%;
   height: 100%;
+  position: relative;
   display: flex;
   flex-direction: row;
   align-items: stretch;
@@ -57,7 +58,7 @@
 
 .Toolbar {
   height: 2.25rem;
-  padding: 0 0.5rem;
+  padding: 0 0.25rem;
   flex: 0 0 auto;
   display: flex;
   align-items: center;
