+++ src/devtools/views/Settings/SettingsModal.css
@@ -1,0 +1,43 @@
+.Background {
+  position: absolute;
+  width: 100%;
+  top: 0;
+  bottom: 0;
+  background-color: var(--color-modal-background);
+  display: flex;
+  align-items: flex-start;
+  justify-content: center;
+  font-size: var(--font-size-sans-normal);
+  padding: 1rem;
+}
+
+.Modal {
+  display: flex;
+  flex-direction: column;
+  flex: 0 1 auto;
+  max-height: 100%;
+  background-color: var(--color-background);
+  border: 1px solid var(--color-border);
+  border-radius: 0.25rem;
+  overflow: auto;
+  width: 400px;
+  max-width: 100%;
+}
+
+.Spacer {
+  flex: 1;
+}
+
+.Tabs {
+  display: flex;
+  flex-direction: row;
+  border-bottom: 1px solid var(--color-border);
+  padding-right: 0.25rem;
+  flex: 0 0 auto;
+}
+
+.Content {
+  padding: 0.5rem;
+  flex: 0 1 auto;
+  overflow: auto;
+}
